Roll controller changes

+++ server/Controllers/RollController.cs
@@ -17,13 +17,23 @@ public class RollController : ControllerBase
     }
 
     [HttpPost("{characterId}/{rollType}/{rollTypeId}")]
-    public async Task<ActionResult<Roll>> Post(long characterId, int? rollType, long? rollTypeId, [FromBody] Roll roll)
+    public async Task<ActionResult<Roll>> Post(long? characterId, long? gameMasterId, int? rollType, long? rollTypeId, [FromBody] Roll roll)
     {
         roll.CharacterId = characterId;
+        roll.GameMasterId = gameMasterId;
         roll.RollType = (RollType)rollType;
         roll.RollIndex = 10;
         if (roll.RollType != RollType.Success)
+        {
             roll.RollTypeId = rollTypeId;
+            var rand = new Random();
+
+            roll.NumberOfDice = 3;
+
+            for (int i = 0; i < roll.NumberOfDice; i++)
+                roll.Result += rand.Next(1, 7);
+            roll.Result += roll.Modififer;
+        }
         else 
         {
             roll.RollTypeId = 0;
@@ -35,16 +45,43 @@ public class RollController : ControllerBase
                 roll.Result += rand.Next(1, 7);
             roll.Result += roll.Modififer;
         }
-        foreach (Roll r in _context.Rolls.Where(R => R.CharacterId == characterId))
+        if (characterId > 0 && gameMasterId > 0)
         {
-            r.RollIndex -= 1;
-            _context.Rolls.Update(r);
-
-            if (r.RollIndex < 1)
+            return Conflict("You can not provide IDs for both Character and Game Master when rolling dice.");
+        }
+        if (characterId > 0)
+        {
+            roll.GameMasterId = null;
+            var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == characterId);
+            if (character == null) return NotFound("No Character matches the provided ID.");
+            foreach (Roll r in _context.Rolls.Where(R => R.CharacterId == characterId))
             {
-                _context.Rolls.Remove(r);
+                r.RollIndex -= 1;
+                _context.Rolls.Update(r);
+
+                if (r.RollIndex < 1)
+                {
+                    _context.Rolls.Remove(r);
+                }
             }
         }
+        else if (gameMasterId > 0)
+        {
+            roll.CharacterId = null;
+            var gameMaster = await _context.GameMasters.FirstOrDefaultAsync(gm => gm.Id == gameMasterId);
+            if (gameMaster == null) return NotFound("No Character matches the provided ID.");
+            foreach (Roll r in _context.Rolls.Where(R => R.GameMasterId == gameMasterId))
+            {
+                r.RollIndex -= 1;
+                _context.Rolls.Update(r);
+
+                if (r.RollIndex < 1)
+                {
+                    _context.Rolls.Remove(r);
+                }
+            }
+        }
+
         _context.Rolls.Add(roll);
         await _context.SaveChangesAsync();
